helpful notice in the example when schema validation is on
I downloaded the repo and ran the example in an android emulator in debug mode.  At first I thought wow this is slow, but then dug around a bit and found that schema validation was on.  This is on by default in debug mode which makes sense ... once you know about it.

I worry someone may get the wrong impression from the example ... that things are slow.

This adds an item to the top of the sample list letting you know that things may be slow.